+++ ft_strlcat.c
@@ -6,7 +6,7 @@
 /*   By: muaktas <muaktas@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/10 14:20:27 by mustafa           #+#    #+#             */
-/*   Updated: 2026/01/11 02:59:08 by muaktas          ###   ########.fr       */
+/*   Updated: 2026/01/13 14:10:35 by muaktas          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 
 size_t	ft_strlcat(char *dst, const char *src, size_t size)
 {
-	int	i;
-	int	j;
-	int	length_of_src;
-	int	init_length_of_dst;
+	size_t	i;
+	size_t	j;
+	size_t	length_of_src;
+	size_t	init_length_of_dst;
 
 	length_of_src = ft_strlen(src);
 	init_length_of_dst = ft_strlen(dst);
